MT51053: Move public/statistiques/export.php to StatisticController
- Move public/statistiques/export.php to StatisticController
- Delete public/statistiques/export.php
- Generate CSV/XLS export using Symfony Response instead of direct file output
- Use entitymanager
- Remove AJAX export, use Symfony routes
- Removed deprecated export_stat() JavaScript function
- Replaced all javascript:export_stat(...) links with Symfony path() links
- Update header to application/octet-stream
- Removed obsolete public/data
- Move verif_select() to statistics.js and include it in statistics templates if it isn't included
- Use Symfony session for stat_debut, stat_fin and stat_tab
