+++ .gitignore
@@ -4,10 +4,7 @@
 /.php_cs.cache
 
 custom_options.php
-public/data/*
-!public/data/index.php
-public/include/test.php
-public/lang/custom.php
+legacy/Lang/custom.php
 
 ###> themes ###
 public/themes/*
