+++ config/permissions.yaml
@@ -301,6 +301,8 @@
     - 17
 "statistics.attendeesmissing":
     - 1301
+"statistics.export":
+    - 17
 "statistics.index":
     - 17
 "statistics.position":
