--- public/data/index.php
@@ -1,15 +1,0 @@
-<?php
-/**
-Planning Biblio, Version 1.9.4
-Licence GNU/GPL (version 2 et au dela)
-Voir les fichiers README.md et LICENSE
-@copyright 2011-2018 Jérôme Combes
-
-Fichier : data/index.php
-Création : 8 décembre 2014
-Dernière modification : 2 avril 2015
-Auteur : Jérôme Combes, jerome@planningbilbio.fr
-
-Description :
-Fichier permettant de conserver le dossier data dans git
-*/
