+++ templates/statistics/positionsummary.html.twig
@@ -69,8 +69,8 @@
               <tr>
                 <td>Exporter </td>
                 <td>
-                  <a href='javascript:export_stat("postes_synthese","csv");'>CSV</a>&nbsp;&nbsp;
-                  <a href='javascript:export_stat("postes_synthese","xls");'>XLS</a>
+                  <a href="{{ path('statistics.export', {name: 'postes_synthese', type: 'csv'}) }}">CSV</a>&nbsp;&nbsp;
+                  <a href="{{ path('statistics.export', {name: 'postes_synthese', type: 'xls'}) }}">XLS</a>
                 </td>
               </tr>
             </table>
@@ -119,4 +119,5 @@
 {% endblock %}
 {% block specificjs %}
   <script type='text/JavaScript'>document.form.tri.value='{{ tri }}';</script>
+  <script type="text/JavaScript" src="{{ asset('js/statistic.js') }}"></script>
 {% endblock %}
