+++ templates/statistics/saturday.html.twig
@@ -86,8 +86,8 @@
               <tr>
                   <td>Exporter </td>
                   <td>
-                  <a href='javascript:export_stat("samedis","csv");'>CSV</a>&nbsp;&nbsp;
-                  <a href='javascript:export_stat("samedis","xsl");'>XLS</a>
+                  <a href="{{ path('statistics.export', {name: 'samedis', type: 'csv'}) }}">CSV</a>&nbsp;&nbsp;
+                  <a href="{{ path('statistics.export', {name: 'samedis', type: 'xsl'}) }}">XLS</a>
                   </td>
               </tr>
               <tr>
@@ -182,5 +182,5 @@
 {% endblock %}
 
 {% block specificjs %}
-  <script type="text/JavaScript" src="{{ asset('js/statistics.js') }}"></script>
+  <script type="text/JavaScript" src="{{ asset('js/statistic.js') }}"></script>
 {% endblock %}
