+++ templates/statistics/service.html.twig
@@ -64,8 +64,8 @@
               <tr>
                 <td>Exporter </td>
                 <td>
-                  <a href='javascript:export_stat("service","csv");'>CSV</a>&nbsp;&nbsp;
-                  <a href='javascript:export_stat("service","xsl");'>XLS</a>
+                  <a href="{{ path('statistics.export', {name: 'service', type: 'csv'}) }}">CSV</a>&nbsp;&nbsp;
+                  <a href="{{ path('statistics.export', {name: 'service', type: 'xsl'}) }}">XLS</a>
                 </td>
               </tr>
               <tr>
@@ -206,5 +206,5 @@
 {% endblock %}
 
 {% block specificjs %}
-  <script type="text/JavaScript" src="{{ asset('js/statistics.js') }}"></script>
+  <script type="text/JavaScript" src="{{ asset('js/statistic.js') }}"></script>
 {% endblock %}
